+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## future release
 
 * Log attempts to reconnect to the MQTT broker (https://github.com/Jalle19/eda-modbus-bridge/issues/36)
+* Add documentation on how to physically connect to the unit
 
 ## 2.1.0