+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Log attempts to reconnect to the MQTT broker (https://github.com/Jalle19/eda-modbus-bridge/issues/36)
 * Add documentation on how to physically connect to the unit
+* Expose the device state (https://github.com/Jalle19/eda-modbus-bridge/issues/46)
 
 ## 2.1.0